power-threejs-game-builder/infra/lambdas/preprocess/requirements.txt
rembg>=2.0.50
Pillow>=10.0
boto3>=1.34.0
